requirements.txt
PyQt6==6.5.1
python-dotenv==1.0.0
requests==2.31.0
